add kill log at the end of service

+++ src/main.c
@@ -26,12 +26,22 @@ int main(int argc, char **argv) {
         }
     }
 
-    char *cFinderPid = (char *) malloc(sizeof(int));
-    sprintf(cFinderPid, "%d", finder_pid);
-    if (execl("./server", "server", cFinderPid, argv[1], NULL) == -1) {
-        perror("main: error on server fork");
-        kill(finder_pid, SIGKILL);
-        exit(EXIT_FAILURE);
+    int server_pid;
+    if ((server_pid = fork()) == 0) {
+        char *cFinderPid = (char *) malloc(sizeof(int));
+        sprintf(cFinderPid, "%d", finder_pid);
+        if (execl("./server", "server", cFinderPid, argv[1], NULL) == -1) {
+            perror("main: error on server fork");
+            kill(finder_pid, SIGKILL);
+            kill(log_pid, SIGKILL);
+            exit(EXIT_FAILURE);
+        }
+    }
+
+    waitpid(server_pid, NULL, 0);
+
+    if(kill(log_pid, SIGKILL)==-1){
+        perror("main: kill log");
     }
 
     return EXIT_SUCCESS;
